requirements.txt
torch>=2.2
torchvision>=0.17
segmentation-models-pytorch>=0.3.3
albumentations>=1.4.7
opencv-python-headless>=4.10.0
numpy>=1.26
pandas>=2.2
rasterio>=1.3
geopandas>=1.0
shapely>=2.0
scikit-image>=0.24
tqdm>=4.66
pyyaml>=6.0
matplotlib>=3.8.0